Folder Structure
Made crucial changes to folder structure
