--- src/Food.java
@@ -1,9 +1,0 @@
-
-public abstract class Food extends Item{
-	
-	public Food(int x, int y){
-		super(x, y);
-	}
-	
-	public abstract int getHealthPoints();
-}
