--- src/Item.java
@@ -1,12 +1,0 @@
-
-public abstract class Item extends GameObject{
-	
-	public Item(int x, int y){
-		super(x, y);
-	}
-	
-	public abstract int getStoreSellingPrice();
-	
-	
-
-}
