+++ src/source/model/Food.java
@@ -1,0 +1,9 @@
+
+public abstract class Food extends Item{
+	
+	public Food(int x, int y){
+		super(x, y);
+	}
+	
+	public abstract int getHealthPoints();
+}
