+++ src/source/model/Item.java
@@ -1,0 +1,12 @@
+
+public abstract class Item extends GameObject{
+	
+	public Item(int x, int y){
+		super(x, y);
+	}
+	
+	public abstract int getStoreSellingPrice();
+	
+	
+
+}
