+++ src/source/view/InventoryPanel.java
